+++ src/calendar/day/basic/driver.js
@@ -1,6 +1,6 @@
 import {ComponentDriver} from 'react-component-driver';
 import Day from '.';
-import {extractStyles} from '../../../../test';
+import {extractStyles} from '../../../../testUtils';
 
 export class BasicDayDriver extends ComponentDriver {
   constructor() {
