+++ tsconfig.json
@@ -22,5 +22,5 @@
   },
   // "include": ["src/calendar/day/basic", "src/global.d.ts"],
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "test"]
+  "exclude": ["node_modules", "testUtils"]
 }
